move defineConfig to check-sql.config.ts
