+++ packages/eslint-plugin/src/index.ts
@@ -1,5 +1,5 @@
 import "source-map-support/register";
 import rules from "./rules";
-import { defineConfig } from "./rules/check-sql.rule";
+import { defineConfig } from "./rules/check-sql.config";
 
 export = { rules, defineConfig };
